Added Update Task on Frontend and Backend as well as changed Frontend components structure
