+++ fe/FE_src/src/components/Admin/Pages/MgtGroup.svelte
@@ -1,0 +1,14 @@
+<script>
+  import Navbar from "../NavBar/IsLoggedInAdmin.svelte";
+  import GroupsTable from "./TableGroups.svelte";
+</script>
+
+<style>
+</style>
+
+<Navbar />
+
+<div class="container-fluid">
+  <br />
+  <GroupsTable />
+</div>
