+++ fe/FE_src/src/components/Kanban/Pages/Dashboard.svelte
@@ -1,0 +1,17 @@
+<script>
+    import axios from "axios";
+    import { errorToast, successToast } from "../../toast";
+    import { Form, FormGroup, Input, Label } from "sveltestrap";
+</script>
+  
+<style>
+</style>
+
+<!-- TO BE DONE BY BEATRICE & JIE WEI & DARREL & HUI EN -->
+<!-- This is dashboard where plan(s) and task(s) will be displayed after clicking into an application -->
+<!-- 1. App Info -->
+<!-- 2. Add Plan -->
+<!-- 3. Add Task -->
+<!-- 4. Update Task -->
+<!-- 5. Move Task -->
+<!-- 6. Display Plan and Task -->
