--- fe/FE_src/src/components/Pages/MgtGroup.svelte
@@ -1,15 +1,0 @@
-<script>
-  import Navbar from "../Navbar/IsLoggedInAdmin.svelte";
-  import AddUserToGroup from "./AddUserToGroup.svelte";
-  import TableGroup from "./TableGroup.svelte";
-</script>
-
-<style>
-</style>
-
-<Navbar />
-
-<div class="container-fluid">
-  <br />
-  <TableGroup />
-</div>
